Factor Out Remote Copasi Call
