+++ vcell-core/src/main/java/org/vcell/optimization/ProfileLikelihood.java
@@ -1,4 +1,4 @@
-package copasi.modelopt;
+package org.vcell.optimization;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.Arrays;
